Create Lab-#9-exercises-1-3.js
We were instructed to create or reuse a sprite of our choosing and make it so that the sprite would bounce off the x and y of the canvas. We were also instructed to make it so that 50 of them would appear and bounce around at different speeds and positions randomly.